+++ packages/blip/app/components/chart/daily.js
@@ -369,6 +369,7 @@ class Daily extends React.Component {
                 chartType={this.chartType}
                 dataUtil={this.props.dataUtil}
                 endpoints={endpoints}
+                loading={loading}
               />
             </div>
           </div>